Deleting a device that is not in the database now gives an error.

+++ source/database/DeviceDatabase.py
@@ -19,7 +19,7 @@ class DeviceDatabase:
 
     def get_device(self, device_id):
         """Instantiates the device with the given device_id"""
-        data = self._devices.find_one({"_id" : device_id})
+        data = self._devices.find_one({"_id": device_id})
         device = self._get_class(data["module"], data["class"])
         return device(self, device_id)
 
@@ -29,25 +29,26 @@ class DeviceDatabase:
         self._devices.insert_one(plugin)
 
     def delete_device(self, device_id):
-        self._devices.delete_one({"_id" : device_id})
+        if self._devices.find({"_id": device_id}).count() > 0:
+            self._devices.delete_one({"_id": device_id})
+        else:
+            raise Exception("No device with this id")
 
     def update_field(self, device_id, field, new_value):
-        self._devices.find_one_and_update({"_id" : device_id}, {"$set": {field: new_value}})
+        self._devices.find_one_and_update({"_id": device_id}, {"$set": {field: new_value}})
 
     def get_field(self, device_id, field):
-        data = self._devices.find_one({"_id" : device_id})
+        data = self._devices.find_one({"_id": device_id})
         return data[field]
 
     def get_activator_field(self, device_id, activator_id, field):
-        data = self._devices.find_one({"_id" : device_id})
+        data = self._devices.find_one({"_id": device_id})
         return data["activators"][activator_id][field]
 
     def update_activator_field(self, device_id, activator_id, field, new_value):
-        self._devices.find_one_and_update({"_id" : device_id}
-                                          , {"$set": {"activators."+activator_id+"."+field: new_value}})
+        self._devices.find_one_and_update({"_id": device_id}
+                                          , {"$set": {"activators." + activator_id + "." + field: new_value}})
 
     def _get_class(self, module, class_name):
         module = importlib.import_module(module)
         return getattr(module, class_name)
-
-
